requirements.txt
# Требуемая версия Python: 3.9+
# Установка зависимостей: pip install -r requirements.txt

# Зависимости для совместимости с новыми версиями Python
setuptools>=65.0.0
wheel

# Основные библиотеки для работы с данными
numpy==1.26.4
pandas==2.2.1
scipy==1.13.0

# Библиотеки для машинного обучения
scikit-learn==1.5.2
catboost==1.2.5

# Библиотеки для глубокого обучения
torch==2.2.1
torchvision==0.17.1

# Библиотеки для работы с текстом и NLP
transformers==4.39.0
sentence-transformers==2.7.0
huggingface-hub==0.22.0

# Библиотеки для визуализации
matplotlib==3.8.3
seaborn==0.13.2

# Библиотеки для обработки изображений
Pillow==10.2.0
img2vec-pytorch==1.0.2

# Библиотеки для веб-скрапинга и обработки HTML
beautifulsoup4==4.12.3
lxml==5.1.0

# Библиотеки для параллельных вычислений
joblib==1.3.2

# Библиотеки для веб-разработки
fastapi==0.110.0
uvicorn==0.27.0

# Библиотеки для работы с конфигурациями
pydantic==2.6.0

# Библиотеки для генерации отчетов
ydata-profiling==4.18.1
jupyter==1.0.0
notebook==7.1.0

# Прочие полезные библиотеки
tqdm
pyhumps

playwright==1.40.0
aiohttp==3.9.0
openai==1.6.1